Do not restrict to TOKENIZED_CARD
With signature verification and message expiration check, this limitation from the inital scrappy implementation can be removed.

This also means the consuming application has full access to details like paymentMethod, which can be used for example with ActiveMerchant to build either a ActiveMerchant::Billing::NetworkTokenizationCreditCard or a ActiveMerchant::Billing::CreditCard object.